feat(infrastructure) : 메인 페이지 로그인 API 로직 구현

+++ src/main/java/com/neighbors/tohero/application/mainPage/dto/GetMainPageInfoResponse.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public record GetMainPageInfoResponse(
         List<TopNotices> topNotices,
-        int writtenLetterNumber,
+        long writtenLetterNumber,
         List<OpenedLetter> openedLetters
 ) {
     public record TopNotices(
